+++ src/bika/lims/browser/dashboard/dashboard.py
@@ -19,5 +19,5 @@
 # Some rights reserved, see README and LICENSE.
 
 # BBB
-from senaite.core.browser.dashboard.dashboard import DashboardView  # noqa
-from senaite.core.browser.dashboard.dashboard import setup_dashboard_panels_visibility_registry  # noqa
+from senaite.core.browser.dashboard.dashboard import DashboardView
+from senaite.core.browser.dashboard.dashboard import setup_dashboard_panels_visibility_registry