+++ src/bika/lims/catalog/report_catalog.py
@@ -6,7 +6,7 @@ from App.class_init import InitializeClass
 from bika.lims.catalog.base import BaseCatalog
 from bika.lims.interfaces import IBikaCatalogReport
 from zope.interface import implements
-from senaite.core.catalog import REPORT_CATALOG as CATALOG_REPORT_LISTING  # noqa
+from senaite.core.catalog import REPORT_CATALOG as CATALOG_REPORT_LISTING
 
 
 class BikaCatalogReport(BaseCatalog):