+++ src/senaite/core/events/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from senaite.core.events.upgrade import BeforeUpgradeStepEvent  # noqa
-from senaite.core.events.upgrade import AfterUpgradeStepEvent  # noqa
+from senaite.core.events.upgrade import BeforeUpgradeStepEvent
+from senaite.core.events.upgrade import AfterUpgradeStepEvent